pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "order-analysis-langgraph"
version = "0.1.0"
description = "LangGraph Studio application for Order Analysis Agent testing"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core LangGraph dependencies
    "langgraph>=0.4.8",
    "langchain>=0.3.26",
    "langchain-core>=0.3.66",
    "langchain-google-vertexai>=2.0.26",
    "langsmith>=0.4.1",
    "langgraph-cli[inmem]>=0.1.59",
    # Google Cloud dependencies
    "google-cloud-bigquery>=3.27.0",
    "google-auth>=2.37.0",
    # Data processing
    "pandas>=2.2.3",
    "numpy>=2.2.3",
    "db-dtypes>=1.3.0",
    # Configuration and utilities
    "pydantic>=2.11.7",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.1",
    "structlog>=25.4.0",
    # Development and testing
    "python-dateutil>=2.9.0.post0",
    "pytz>=2025.2",
    "pyyaml>=6.0.2",
    "jinja2>=3.1.5",
    "aiofiles>=24.1.0",
    "httpx>=0.28.1",
    "tenacity>=8.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.setuptools]
py-modules = []

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
